# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API Endpoint for handling organizational financial reimbursement requests",
        "title": "Financial Request Handler API",
        "contact": {
            "name": "APP Support",
            "email": "app_support@norainc.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "either logs a user in or creates an account depending on previous use of the application",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "no-cache",
                    "auth"
                ],
                "summary": "login user",
                "operationId": "login-user",
                "parameters": [
                    {
                        "description": "user's account information",
                        "name": "login-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.NilRes"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.NilRes"
                        }
                    }
                }
            }
        },
...
        "/check": {
            "put": {
                "description": "allows a logged in user to edit their pending or reject check requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "check",
                    "no-cache"
                ],
                "summary": "edit check request",
                "operationId": "edit-check",
                "parameters": [
                    {
                        "description": "check request information to update",
                        "name": "check-request-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EditCheckInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.CheckOverviewRes"
                        }
                    }
                }
            },
            "post": {
                "description": "creates a check request for a logged in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "check",
                    "no-cache"
                ],
                "summary": "add check request",
                "operationId": "add-check",
                "parameters": [
                    {
                        "description": "new check request information",
                        "name": "check-request-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CheckRequestInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.CheckOverviewRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "allows a logged in user to delete and removes one of their check requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "check",
                    "no-cache"
                ],
                "summary": "delete check request",
                "operationId": "delete-check",
                "parameters": [
                    {
                        "description": "check request id to delete",
                        "name": "check-request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FindCheckInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.CheckOverviewRes"
                        }
                    }
                }
            }
        },
        "/check/approve": {
            "post": {
                "description": "allows an administrative user to approve a specific check request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "check",
                    "no-cache",
                    "admin"
                ],
                "summary": "approve check request",
                "operationId": "approve-check",
                "parameters": [
                    {
                        "description": "check request id to approve",
                        "name": "request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ApproveRejectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.CheckOverviewRes"
                        }
                    }
                }
            }
        },
        "/check/detail": {
            "get": {
                "description": "generates detailed information for a specific check request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "check"
                ],
                "summary": "check request detail",
                "operationId": "check-detail",
                "parameters": [
                    {
                        "description": "check request id to find",
                        "name": "check-request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FindCheckInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.CheckDetailRes"
                        }
                    }
                }
            }
        },
        "/check/monthly": {
            "get": {
                "description": "creates a monthly check report for the entire organization for a given month and year",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "check",
                    "reports",
                    "admin"
                ],
                "summary": "monthly check request report",
                "operationId": "monthly-check",
                "parameters": [
                    {
                        "description": "month and year for report on organization wide check requests",
                        "name": "month-year-input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MonthlyRequestInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.CheckOverviewsRes"
                        }
                    }
                }
            }
        },
        "/check/reject": {
            "post": {
                "description": "allows a logged in administrative user to reject a check request until further edits have been made",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "check",
                    "no-cache",
                    "admin"
                ],
                "summary": "reject check request",
                "operationId": "reject-check",
                "parameters": [
                    {
                        "description": "check request id to reject",
                        "name": "request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ApproveRejectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.CheckOverviewRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Action": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "models.ApproveRejectRequest": {
            "type": "object",
            "required": [
                "request_id"
            ],
            "properties": {
                "request_id": {
                    "type": "string"
                }
            }
        },
        "models.Category": {
            "type": "string",
            "enum": [
                "IOP",
                "INTAKE",
                "PEERS",
                "ACT_TEAM",
                "IHBT",
                "PERKINS",
                "MENS_HOUSE",
                "NEXT_STEP",
                "LORAIN",
                "PREVENTION",
                "ADMINISTRATIVE",
                "FINANCE"
            ],
            "x-enum-varnames": [
                "IOP",
                "INTAKE",
                "PEERS",
                "ACT_TEAM",
                "IHBT",
                "PERKINS",
                "MENS_HOUSE",
                "NEXT_STEP",
                "LORAIN",
                "PREVENTION",
                "ADMINISTRATIVE",
                "FINANCE"
            ]
        },
        "models.CheckRequestInput": {
            "type": "object",
            "required": [
                "category",
                "credit_card",
                "date",
                "description",
                "grant_id",
                "purchases",
                "receipts",
                "vendor"
            ],
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "credit_card": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "grant_id": {
                    "type": "string"
                },
                "purchases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Purchase"
                    }
                },
                "receipts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "vendor": {
                    "$ref": "#/definitions/models.Vendor"
                }
            }
        },
        "models.Check_Request": {
            "type": "object",
            "properties": {
                "action_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Action"
                    }
                },
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "created_at": {
                    "type": "string"
                },
                "credit_card": {
                    "type": "string"
                },
                "current_status": {
                    "type": "string"
                },
                "current_user": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "grant_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_user_before_reject": {
                    "type": "string"
                },
                "order_total": {
                    "type": "number"
                },
                "purchases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Purchase"
                    }
                },
                "receipts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                },
                "vendor": {
                    "$ref": "#/definitions/models.Vendor"
                }
            }
        },
        "models.Check_Request_Overview": {
            "type": "object",
            "properties": {
                "current_status": {
                    "type": "string"
                },
                "current_user": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "order_total": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.EditCheckInput": {
            "type": "object",
            "required": [
                "category",
                "credit_card",
                "date",
                "description",
                "grant_id",
                "id",
                "purchases",
                "receipts",
                "user_id",
                "vendor"
            ],
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "credit_card": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "grant_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "purchases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Purchase"
                    }
                },
                "receipts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                },
                "vendor": {
                    "$ref": "#/definitions/models.Vendor"
                }
            }
        },
...
        "models.FindCheckInput": {
            "type": "object",
            "required": [
                "check_id"
            ],
            "properties": {
                "check_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Month": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12
            ],
            "x-enum-varnames": [
                "January",
                "February",
                "March",
                "April",
                "May",
                "June",
                "July",
                "August",
                "September",
                "October",
                "November",
                "December"
            ]
        },
        "models.MonthlyRequestInput": {
            "type": "object",
            "required": [
                "month",
                "year"
            ],
            "properties": {
                "month": {
                    "$ref": "#/definitions/models.Month"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "models.Purchase": {
            "type": "object",
            "required": [
                "amount",
                "description",
                "grant_line_item"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "grant_line_item": {
                    "type": "string"
                }
            }
        },
        "models.UserLogin": {
            "type": "object",
            "required": [
                "email",
                "id",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Vendor": {
            "type": "object",
            "required": [
                "address_line_one",
                "name"
            ],
            "properties": {
                "address_line_one": {
                    "type": "string"
                },
                "address_line_two": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "responses.CheckDetailRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.Check_Request"
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
        "responses.CheckOverviewRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.Check_Request_Overview"
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
        "responses.CheckOverviewsRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Check_Request_Overview"
                    }
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
...
        "responses.NilRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
...
    }
}